tasks/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
